chore!: move `src/simulat/core/surfaces/game_map` to `src/simulat/core/scenes/game_scene/game_map`
BREAKING CHANGE: Moved `game_map` folder to `src/simulat/core/scenes/game_scene/game_map/`.
