+++ src/simulat/core/scenes/game_scene/game_map/sidebar.py
